requirements.txt
mysql-connector-python==8.0.29
protobuf==3.20.1
tk==0.1.0
